+++ src/lib/DatePickerInput.tsx
@@ -1,4 +1,4 @@
-import { Controller, FieldValues, useFormContext } from "react-hook-form";
+import { Controller, FieldValues } from "react-hook-form";
 import { useSafeNameId } from "src/lib/hooks/useSafeNameId";
 import { FormGroupLayout } from "./FormGroupLayout";
 import { CommonInputProps } from "./types/CommonInputProps";
@@ -6,7 +6,7 @@ import DatePicker, { ReactDatePickerProps } from "react-datepicker";
 import { useCallback, useEffect, useState } from "react";
 import { utcToZonedTime, zonedTimeToUtc } from "date-fns-tz";
 import { getUtcTimeZeroDate } from "./helpers/dateUtils";
-import { useInternalFormContext } from "./context/InternalFormContext";
+import { useFormContext } from "./context/FormContext";
 
 interface DatePickerInputProps<T extends FieldValues> extends Omit<CommonInputProps<T>, "onChange" | "style"> {
   datePickerProps?: Omit<ReactDatePickerProps, "onChange" | "selected" | "id" | "className" | "onBlur">;
@@ -25,8 +25,7 @@ const DEFAULT_DATE_TIME_FORMAT = "dd.MM.yyyy HH:mm";
 
 const DatePickerInput = <T extends FieldValues>(props: DatePickerInputProps<T>) => {
   const { name, id } = useSafeNameId(props.name, props.id);
-  const { control, getValues, setValue } = useFormContext();
-  const { disabled: formDisabled } = useInternalFormContext<T>();
+  const { control, getValues, setValue, disabled: formDisabled } = useFormContext();
 
   const {
     disabled,